# perf: disable router cache refresh that occurs after every page transition or page load